[VT]: Additional VT Server change commands
Implemented change string value command response.
Implemented change fill attribute command response.
Implemented change child position command response.
Added input string value setter and getter.

+++ isobus/src/isobus_language_command_interface.cpp
@@ -119,7 +119,7 @@ namespace isobus
 		};
 		return CANNetworkManager::CANNetwork.send_can_message(static_cast<std::uint32_t>(CANLibParameterGroupNumber::LanguageCommand),
 		                                                      buffer.data(),
-		                                                      buffer.size(),
+		                                                      static_cast<std::uint32_t>(buffer.size()),
 		                                                      myControlFunction,
 		                                                      nullptr);
 	}
